✨(api) add Session and Status dynamic schemas
First draft implementation of database schemas for dynamique data.
