+++ src/api/qualicharge/models/dynamic.py
@@ -3,9 +3,9 @@
 from enum import StrEnum
 from typing import Optional
 
-from pydantic import BaseModel, Field
+from pydantic import PositiveFloat
 from pydantic.types import PastDatetime
-from typing_extensions import Annotated
+from sqlmodel import Field, SQLModel
 
 
 class EtatPDCEnum(StrEnum):
@@ -33,10 +33,9 @@ class EtatPriseEnum(StrEnum):
     INCONNU = "inconnu"
 
 
-class Status(BaseModel):
-    """IRVE dynamic model: point of charge status."""
+class StatusBase(SQLModel):
+    """Base point of charge status."""
 
-    id_pdc_itinerance: Annotated[str, Field(pattern="^[A-Z]{2}[A-Z0-9]{4,33}$")]
     etat_pdc: EtatPDCEnum
     occupation_pdc: OccupationPDCEnum
     horodatage: PastDatetime
@@ -46,10 +45,29 @@ class Status(BaseModel):
     etat_prise_type_ef: Optional[EtatPriseEnum]
 
 
-class Session(BaseModel):
-    """IRVE dynamic model: point of charge sessions."""
+class StatusCreate(StatusBase):
+    """Point of charge status create."""
 
-    id_pdc_itinerance: Annotated[str, Field(pattern="^[A-Z]{2}[A-Z0-9]{4,33}$")]
+    id_pdc_itinerance: str = Field(
+        regex="(?:(?:^|,)(^[A-Z]{2}[A-Z0-9]{4,33}$|Non concerné))+$"
+    )
+
+
+class StatusRead(StatusCreate):
+    """Point of charge status read."""
+
+
+class SessionBase(SQLModel):
+    """Base point of charge sessions."""
+
     start: PastDatetime
     end: PastDatetime
-    energy: float
+    energy: PositiveFloat
+
+
+class SessionCreate(SessionBase):
+    """Point of charge sessions create."""
+
+    id_pdc_itinerance: str = Field(
+        regex="(?:(?:^|,)(^[A-Z]{2}[A-Z0-9]{4,33}$|Non concerné))+$"
+    )
